Move Nyx core programming to nyx package.
