+++ nyx/__init__.py
@@ -1,1 +1,4 @@
+from .nyx import Nyx
+from .nyxcommands import *
 from .nyxhelp import NyxHelpFormatter
+from .nyxsplash import splash
